Add files via upload
First off, we've structured the page using HTML, employing semantic <h1> headings to delineate different sections or pages of our site. Each heading acts as a clickable link, guiding users to various destinations within the site.

To bring our design to life, we've employed CSS to style the page elements. While the specifics of the styling are encapsulated in an external CSS file (i-1.css), rest assured, we're prioritizing a clean and modern aesthetic. The page will exude professionalism while remaining visually engaging.

Now, for a touch of interactivity, we've sprinkled in some JavaScript magic. As users scroll down the page, we've implemented a feature that dynamically changes the background color of a navbar to provide visual feedback and enhance the user experience. It's a subtle yet effective way to keep users engaged as they explore our content.